+++ crates/matrix-sdk/src/authentication/oauth/error.rs
@@ -31,8 +31,6 @@ use ruma::{
 
 #[cfg(feature = "e2e-encryption")]
 pub use super::cross_process::CrossProcessRefreshLockError;
-#[cfg(not(target_arch = "wasm32"))]
-pub use super::registration_store::OAuthRegistrationStoreError;
 
 /// An error when interacting with the OAuth 2.0 authorization server.
 pub type OAuthRequestError<T> =
@@ -258,11 +256,6 @@ pub enum OAuthClientRegistrationError {
     /// Deserialization of the registration response failed.
     #[error("failed to deserialize registration response: {0}")]
     FromJson(serde_json::Error),
-
-    /// Failed to access or store the registration in the store.
-    #[cfg(not(target_arch = "wasm32"))]
-    #[error("failed to use registration store: {0}")]
-    Store(#[from] OAuthRegistrationStoreError),
 }
 
 /// Error response returned by server after requesting an authorization code.