+++ crates/matrix-sdk/src/room/mod.rs
@@ -58,6 +58,7 @@ use matrix_sdk_common::{
     timeout::timeout,
 };
 use mime::Mime;
+use reply::Reply;
 #[cfg(feature = "e2e-encryption")]
 use ruma::events::{
     room::encrypted::OriginalSyncRoomEncryptedEvent, AnySyncMessageLikeEvent, AnySyncTimelineEvent,
@@ -138,7 +139,7 @@ pub use self::{
 #[cfg(doc)]
 use crate::event_cache::EventCache;
 use crate::{
-    attachment::{AttachmentConfig, AttachmentInfo, Reply},
+    attachment::{AttachmentConfig, AttachmentInfo},
     client::WeakClient,
     config::RequestConfig,
     error::{BeaconError, WrongRoomState},
@@ -209,7 +210,10 @@ impl Room {
     pub async fn leave(&self) -> Result<()> {
         let state = self.state();
         if state == RoomState::Left {
-            return Err(Error::WrongRoomState(WrongRoomState::new("Joined or Invited", state)));
+            return Err(Error::WrongRoomState(Box::new(WrongRoomState::new(
+                "Joined or Invited",
+                state,
+            ))));
         }
 
         let request = leave_room::v3::Request::new(self.inner.room_id().to_owned());
@@ -225,7 +229,10 @@ impl Room {
     pub async fn join(&self) -> Result<()> {
         let state = self.state();
         if state == RoomState::Joined {
-            return Err(Error::WrongRoomState(WrongRoomState::new("Invited or Left", state)));
+            return Err(Error::WrongRoomState(Box::new(WrongRoomState::new(
+                "Invited or Left",
+                state,
+            ))));
         }
 
         let prev_room_state = self.inner.state();
@@ -2272,9 +2279,7 @@ impl Room {
         if let Some(reply) = reply {
             // Since we just created the event, there is no relation attached to it. Thus,
             // it is safe to add the reply relation without overriding anything.
-            content = self
-                .make_reply_event(content.into(), &reply.event_id, reply.enforce_thread)
-                .await?;
+            content = self.make_reply_event(content.into(), reply).await?;
         }
         Ok(content)
     }
@@ -2939,7 +2944,7 @@ impl Room {
     pub async fn invite_details(&self) -> Result<Invite> {
         let state = self.state();
         if state != RoomState::Invited {
-            return Err(Error::WrongRoomState(WrongRoomState::new("Invited", state)));
+            return Err(Error::WrongRoomState(Box::new(WrongRoomState::new("Invited", state))));
         }
 
         let invitee = self
@@ -2990,7 +2995,10 @@ impl Room {
         let state = self.state();
         match state {
             RoomState::Joined | RoomState::Invited | RoomState::Knocked => {
-                return Err(Error::WrongRoomState(WrongRoomState::new("Left / Banned", state)));
+                return Err(Error::WrongRoomState(Box::new(WrongRoomState::new(
+                    "Left / Banned",
+                    state,
+                ))));
             }
             RoomState::Left | RoomState::Banned => {}
         }
@@ -3017,7 +3025,7 @@ impl Room {
         if state == RoomState::Joined {
             Ok(())
         } else {
-            Err(Error::WrongRoomState(WrongRoomState::new("Joined", state)))
+            Err(Error::WrongRoomState(Box::new(WrongRoomState::new("Joined", state))))
         }
     }
 
@@ -3100,7 +3108,7 @@ impl Room {
     ) -> Result<report_content::v3::Response> {
         let state = self.state();
         if state != RoomState::Joined {
-            return Err(Error::WrongRoomState(WrongRoomState::new("Joined", state)));
+            return Err(Error::WrongRoomState(Box::new(WrongRoomState::new("Joined", state))));
         }
 
         let request = report_content::v3::Request::new(