+++ crates/matrix-sdk/src/widget/matrix.rs
@@ -55,7 +55,11 @@ impl MatrixDriver {
     /// Requests an OpenID token for the current user.
     pub(crate) async fn get_open_id(&self) -> Result<OpenIdResponse> {
         let user_id = self.room.own_user_id().to_owned();
-        self.room.client.send(OpenIdRequest::new(user_id)).await.map_err(Error::Http)
+        self.room
+            .client
+            .send(OpenIdRequest::new(user_id))
+            .await
+            .map_err(|error| Error::Http(Box::new(error)))
     }
 
     /// Reads the latest `limit` events of a given `event_type` from the room.
@@ -172,7 +176,7 @@ impl MatrixDriver {
         action: UpdateAction,
     ) -> Result<delayed_events::update_delayed_event::unstable::Response> {
         let r = delayed_events::update_delayed_event::unstable::Request::new(delay_id, action);
-        self.room.client.send(r).await.map_err(Error::Http)
+        self.room.client.send(r).await.map_err(|error| Error::Http(Box::new(error)))
     }
 
     /// Starts forwarding new room events. Once the returned `EventReceiver`